requirements.txt
fastapi[all]==0.104.1
uvicorn[standard]==0.30.6
qdrant-client==1.11.1
pypdf==4.3.1
numpy==1.26.4
requests==2.32.3
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
pydantic==2.5.0
pydantic-settings==2.2.1
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
pytest-mock==3.12.0
mypy==1.11.2
mutmut==2.4.4